Refactor OAD code from GattClient into OADProfile
